+++ src/api/v1/apps/orders/router.py
@@ -3,9 +3,11 @@ from fastapi import APIRouter, status, Request, Depends, HTTPException
 from fastapi.responses import Response, HTMLResponse
 from fastapi.templating import Jinja2Templates
 from sqlmodel import Session
+from typing import List
 from src.api.v1.apps.orders.schemas import ProductCreate, ProductRead, ProductUpdate
-from src.api.v1.apps.orders.services import FlavorService, ProductService
+from src.api.v1.apps.orders.services import FlavorService, MeatService, ProductService
 from src.api.v1.apps.orders.schemas import FlavorCatalogueCreate, FlavorCatalogueRead, FlavorCatalogueUpdate
+from src.api.v1.apps.orders.schemas import MeatCatalogueCreate, MeatCatalogueRead, MeatCatalogueUpdate
 from src.api.v1.apps.orders.models import Product
 from src.api.v1.auth.auth import get_current_user_from_cookie
 from src.config import settings
@@ -217,3 +219,117 @@ async def get_flavor_by_id(
     return flavor
 
 ######### --- Flavors endpoints end --- ##############
+
+
+
+##########  --- Meat catalogue endpoints init --- ##############
+
+# ui template endpoint for meat catalogue management
+@router.get("/flavors/meat/add/ui", response_class=HTMLResponse, status_code=status.HTTP_200_OK)
+async def get_meat_add_page(request: Request,
+                           current_user: dict = Depends(get_current_user_from_cookie),
+                           session: Session = Depends(get_session)
+                           ) -> Response:
+    return templates.TemplateResponse(
+        request=request,
+        name="meat/create.html",
+        context={
+            "config": settings,
+            "user": current_user
+        }
+    )
+
+@router.post("/flavors/meat/add", response_model=MeatCatalogueRead, status_code=status.HTTP_201_CREATED)
+def create_new_meat(meat_in: MeatCatalogueCreate, 
+                    current_user: dict = Depends(get_current_user_from_cookie),
+                    session: Session = Depends(get_session)):
+    
+    meat_service = MeatService(session)
+    return meat_service.create_meat(meat_in=meat_in)
+
+# ui template endpoint for meat catalogue management
+@router.get("/flavors/meat/list", response_class=HTMLResponse)
+async def get_meat_management_page(request: Request,
+                        current_user: dict = Depends(get_current_user_from_cookie),
+                        session: Session = Depends(get_session)
+                        ) -> Response:
+    meat_service = MeatService(session)
+    meat_list = meat_service.get_meat_catalogue()
+    print(meat_list)
+    return templates.TemplateResponse(
+        request=request,
+        name="meat/list.html",
+        context={
+            "config": settings,
+            "meat_list": meat_list,
+            "user": current_user # for nav_bar
+        }
+    )
+
+# general endpoint go first, then the specifics
+@router.get("/flavors/meat/all", response_model=List[MeatCatalogueRead], status_code=status.HTTP_200_OK)
+async def get_meat_all(
+    current_user: dict = Depends(get_current_user_from_cookie),
+    session: Session = Depends(get_session)
+) -> List[MeatCatalogueRead]:
+    """
+    Returns all meat in the catalogue as JSON
+    """
+    meat_service = MeatService(session)
+    meat_list = meat_service.get_meat_catalogue()
+
+    # If the service returns None or an empty list, raise the 404
+    if not meat_list:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Meat catalogue is empty."
+        ) 
+
+    return meat_list
+
+
+@router.get("/flavors/meat/{id}", 
+            response_model=MeatCatalogueRead, 
+            status_code=status.HTTP_200_OK)
+async def get_meat_by_id(
+    id: int,
+    request: Request,
+    current_user: dict = Depends(get_current_user_from_cookie),
+    session: Session = Depends(get_session)
+    ) -> MeatCatalogueRead | None:
+    """
+    Returns an individual meat information
+    """
+
+    meat_service = MeatService(session)
+    meat = meat_service.get_meat_by_id(meat_id=id)
+
+    if not meat:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Meat with ID index {id} could not be located."
+        ) 
+
+    return meat
+
+@router.patch("/flavors/meat/{meat_id}", 
+              response_model=MeatCatalogueRead, status_code=status.HTTP_200_OK)
+async def update_meat(meat_update: MeatCatalogueUpdate,
+                      meat_id: int, 
+                      current_user: dict = Depends(get_current_user_from_cookie),
+                      session: Session = Depends(get_session)
+                         ):
+    
+    meat_service = MeatService(session)
+    
+    return meat_service.update_meat(meat_id=meat_id, meat_in=meat_update)
+
+@router.delete("/flavors/meat/{meat_id}")
+async def delete_meat(meat_id: int, 
+                       current_user: dict = Depends(get_current_user_from_cookie),
+                       session: Session = Depends(get_session)
+                         ):
+    
+    meat_service = MeatService(session)
+    
+    return meat_service.delete_meat(meat_id=meat_id)